More code review, e.g. use Tcl_SetObjResult in stead of Tcl_SetResult, preventing a (char *) type case. No functional changes.

+++ generic/tclEnv.c
@@ -19,7 +19,7 @@ TCL_DECLARE_MUTEX(envMutex)	/* To serialize access to environ. */
 
 
 /* MODULE_SCOPE */
-unsigned long TclEnvEpoch = 0;	/* Epoch of the tcl environment 
+size_t TclEnvEpoch = 0;	/* Epoch of the tcl environment
 				 * (if changed with tcl-env). */
 
 static struct {
